move the logic to the lib folder, create game_status and game_utils files
